ai-in-action/1-invoice-ai-agents/requirements.txt
python-dotenv
typing
langgraph
langchain-openai
langchain-core
pydantic
colorama